# specs/001-tv-channel-programming/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SmartTunarr API
  version: 1.0.0
  description: Intelligent TV Channel Programming System API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /services/{serviceType}:
    get:
      tags: [Services]
      summary: Get service configuration
      operationId: getService
      parameters:
        - $ref: '#/components/parameters/ServiceType'
      responses:
        '200':
          description: Service configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Services]
      summary: Update service configuration
      operationId: updateService
      parameters:
        - $ref: '#/components/parameters/ServiceType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Service updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
  /services/{serviceType}/test:
    post:
      tags: [Services]
      summary: Test service connection
      operationId: testService
      parameters:
        - $ref: '#/components/parameters/ServiceType'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTestResult'
...
components:
  parameters:
...
    ServiceType:
      name: serviceType
      in: path
      required: true
      schema:
        type: string
        enum: [plex, tmdb, tunarr, ollama]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Service:
      type: object
      properties:
        type:
          type: string
          enum: [plex, tmdb, tunarr, ollama]
        name:
          type: string
        url:
          type: string
        is_active:
          type: boolean
        is_configured:
          type: boolean
        last_test:
          type: string
          format: date-time
        last_test_success:
          type: boolean
        default_model:
          type: string
    ServiceUpdate:
      type: object
      properties:
        url:
          type: string
        api_key:
          type: string
        token:
          type: string
        username:
          type: string
        password:
          type: string
        default_model:
          type: string
        is_active:
          type: boolean
    ServiceTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        details:
          type: object
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object